Refactor on BytesConverter pass (#1005)
* refactor

* case typo in import

* comments

* comments on replaceIndexAccessBytesConverter

* changes in utils

* lint

* changes in updateTypeStringBytesConverter

* `selectWarplibFunction` receive `typeNode` arguments
